requirements.txt
aiohttp==3.8.4
aiolimiter==1.1.0
aiosignal==1.3.1
appnope==0.1.3
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
black==23.3.0
blis==0.7.9
bs4==0.0.1
cachetools==5.3.1
catalogue==2.0.8
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
cmake==3.26.1
confection==0.0.4
cymem==2.0.7
decorator==5.1.1
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.5.0/en_core_web_lg-3.5.0-py3-none-any.whl#sha256=c8ac64840c1eb3e3ca7bd38bd1e1c48fb0faeb2449d54d01d5ce629af4595775
executing==1.2.0
filelock==3.10.0
frozenlist==1.3.3
google-api-core==2.11.0
google-auth==2.19.1
google-cloud-core==2.3.2
google-cloud-storage==2.9.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.0
httptools==0.5.0
huggingface-hub==0.13.2
idna==3.4
ipython==8.14.0
jedi==0.18.2
Jinja2==3.1.2
langcodes==3.3.0
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
mpmath==1.3.0
multidict==6.0.4
murmurhash==1.0.9
mypy-extensions==1.0.0
networkx==3.0
numpy==1.24.2
openai==0.27.8
packaging==23.0
pandas==1.5.3
parso==0.8.3
pathspec==0.11.1
pathy==0.10.1
pexpect==4.8.0
phonenumbers==8.13.13
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.5.1
preshed==3.0.8
presidio-analyzer==2.2.33
prompt-toolkit==3.0.38
protobuf==4.21.12
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycryptodome==3.18.0
pydantic==1.10.8
Pygments==2.15.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.2
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
requests-file==1.5.1
rsa==4.9
six==1.16.0
smart-open==6.3.0
soupsieve==2.4.1
spacy==3.5.3
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.6
stack-data==0.6.2
sympy==1.11.1
thinc==8.1.10
tldextract==3.4.4
tokenize-rt==5.0.0
tokenizers==0.13.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.27.1
typer==0.7.0
typing_extensions==4.5.0
urllib3==1.26.15
wasabi==1.1.1
wcwidth==0.2.6
yarl==1.9.2
